+++ src/main/java/no/nav/foreldrepenger/selvbetjening/historikk/Arbeidsgiver.java
@@ -1,5 +1,11 @@
 package no.nav.foreldrepenger.selvbetjening.historikk;
 
-record Arbeidsgiver(String navn, String orgnr) {
+import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.ANY;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+
+@JsonAutoDetect(fieldVisibility = ANY)
+
+public record Arbeidsgiver(String navn, String orgnr) {
 
 }
